requirements.txt
flask
requests
Flask-SQLAlchemy
Flask-Restless
mysql-connector-python
